+++ crates/libs/metadata/src/lib.rs
@@ -6,7 +6,6 @@ mod flags;
 mod imp;
 pub mod reader;
 pub mod writer;
-pub mod writer2;
 
 use bindings::*;
 pub use flags::*;
