+++ crates/libs/metadata/src/writer/imp/blobs.rs
