+++ crates/libs/metadata/src/writer/imp/definitions.rs
