+++ crates/libs/metadata/src/writer/imp/file.rs
