+++ crates/libs/metadata/src/writer/imp/references.rs
