+++ crates/libs/metadata/src/writer/imp/strings.rs
