+++ crates/libs/metadata/src/writer/imp/tables.rs
