--- crates/libs/metadata/src/writer/tables/class_layout.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct ClassLayout {}
