--- crates/libs/metadata/src/writer/tables/constant.rs
@@ -1,7 +1,0 @@
-use super::*;
-
-#[derive(Default)]
-pub struct Constant {
-    pub value: Value,
-    pub(crate) parent_index: HasConstant,
-}
