--- crates/libs/metadata/src/writer/tables/custom_attribute.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct CustomAttribute {}
