--- crates/libs/metadata/src/writer/tables/generic_param.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct GenericParam {}
