--- crates/libs/metadata/src/writer/tables/impl_map.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct ImplMap {}
