--- crates/libs/metadata/src/writer/tables/interface_impl.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct InterfaceImpl {}
