--- crates/libs/metadata/src/writer/tables/member_ref.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct MemberRef {}
