--- crates/libs/metadata/src/writer/tables/module.rs
@@ -1,10 +1,0 @@
-#[derive(Default)]
-pub struct Module {
-    pub name: String,
-}
-
-impl Module {
-    pub fn new(name: &str) -> Self {
-        Self { name: name.to_string() }
-    }
-}
