--- crates/libs/metadata/src/writer/tables/module_ref.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct ModuleRef {}
