--- crates/libs/metadata/src/writer/tables/nested_class.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct NestedClass {}
