--- crates/libs/metadata/src/writer/tables/param.rs
@@ -1,8 +1,0 @@
-use super::*;
-
-#[derive(Default)]
-pub struct Param {
-    pub flags: ParamAttributes,
-    pub sequence: u16,
-    pub name: String,
-}
