--- crates/libs/metadata/src/writer/tables/property.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct Property {}
