--- crates/libs/metadata/src/writer/tables/type_spec.rs
@@ -1,2 +1,0 @@
-#[derive(Default)]
-pub struct TypeSpec {}
