+++ crates/tests/metadata/tests/writer.rs
@@ -1,18 +1,19 @@
 #[test]
 fn writer() {
-    let temp_file = std::env::temp_dir().join("test2.winmd");
+    // ildasm %temp%\test_metadata.winmd
+    let temp_file = std::env::temp_dir().join("test_metadata.winmd");
     {
-        use metadata::writer2::*;
+        use metadata::writer::*;
 
         let mut items = vec![];
 
-        items.push(Struct::new("test_component2", "Inner", vec![Field::new("Value32", Type::F32), Field::new("Value64", Type::F64)]));
+        items.push(Struct::item("test_metadata", "Inner", vec![Field::new("Value32", Type::F32), Field::new("Value64", Type::F64)]));
 
-        items.push(Struct::new("test_component2", "Outer", vec![Field::new("Value", Type::named("test_component2", "Inner"))]));
+        items.push(Struct::item("test_metadata", "Outer", vec![Field::new("Value", Type::named("test_metadata", "Inner"))]));
 
-        items.push(Enum::new("test_component2", "Flags", vec![Constant::new("One", Value::I32(1)), Constant::new("Two", Value::I32(2))]));
+        items.push(Enum::item("test_metadata", "Flags", vec![Constant::new("One", Value::I32(1)), Constant::new("Two", Value::I32(2))]));
 
-        let buffer = write("test_component2", true, &items, &[]);
+        let buffer = write("test_metadata", true, &items, &[]);
         std::fs::write(temp_file, buffer).unwrap();
     }
 }
